+++ package.json
@@ -10,6 +10,7 @@
   "license": "ISC",
   "dependencies": {
     "@typegoose/typegoose": "^11.4.1",
+    "dotenv": "^16.3.1",
     "express": "^4.18.2",
     "jsonwebtoken": "^9.0.1",
     "mongodb": "^5.6.0",
