--- src/controllers/placeholder.controller.ts
@@ -1,16 +1,0 @@
-//! Archivo solo de prueba, editar/borrar en sus branches
-//* Deps
-import { Request, Response } from "express";
-
-//* Models
-//? Agregar modelos
-//? import placeholderModel from '../models/placeholder.model.ts';
-
-export default {
-    getP: (req: Request, res: Response) => {
-        // Codigo
-    },
-    postP: (req: Request, res: Response) => {
-        // Codigo
-    }
-}
