--- src/models/placeholder.model.ts
@@ -1,13 +1,0 @@
-//! Archivo solo de prueba, editar/borrar en sus branches
-//* Deps
-import { getModelForClass, modelOptions, prop } from "@typegoose/typegoose";
-
-@modelOptions({
-    schemaOptions: {"collection": "placeholder"}
-})
-class placeholder {
-    @prop({required: true})
-    test!: string
-}
-
-export const placeholderModel = getModelForClass(placeholder)
