--- src/routes/placeholder.routes.ts
@@ -1,11 +1,0 @@
-//! Archivo solo de prueba, editar/borrar en sus branches
-import express from 'express';
-//? import placeholderController from '../controllers/placeholder.controller.ts'
-
-export const placeholderRouter = express.Router()
-
-placeholderRouter
-    //?.get('/', placeholderController.getP)
-    //?.post('/', placeholderController.postP)
-    //?.patch('/', placeholderController.patchP)
-    //?.delete('/', placeholderController.delP)
